fast-backend/requirements.txt
fastapi
uvicorn[standard]
httpx
psycopg2-binary
sqlalchemy
scikit-learn
pandas
numpy
python-dotenv
